Nishant Portfolio
Developed my portfolio website utilizing basic HTML, CSS, JavaScript, and jQuery. The website features a clean and responsive design, showcasing my projects and skills. Key sections include an introduction, project highlights with links to live demos, and a contact form. Implemented smooth scrolling and hover effects for a better user experience. Used jQuery for DOM manipulation to enhance interactivity, such as toggling the navigation menu on mobile devices. The website is optimized for both desktop and mobile views, ensuring accessibility and ease of navigation. This project serves as a comprehensive representation of my work and capabilities